=tes #17080 Move TestUtils to testkit.SocketUtil
(cherry picked from commit 3df2a822efa7e1330d564ceea7506d606359e17c)

Conflicts:
	akka-actor-tests/src/test/scala/akka/actor/ActorPerfSpec.scala
	akka-actor-tests/src/test/scala/akka/io/UdpIntegrationSpec.scala
	akka-durable-mailboxes/akka-file-mailbox/src/main/scala/akka/actor/mailbox/filebased/filequeue/Journal.scala
	akka-testkit/src/main/scala/akka/testkit/SocketUtil.scala
