+++ akka-actor-tests/src/test/scala/akka/io/UdpIntegrationSpec.scala
@@ -10,7 +10,7 @@ import akka.util.ByteString
 import akka.actor.ActorRef
 import akka.io.Udp._
 import akka.io.Inet._
-import akka.TestUtils._
+import akka.testkit.SocketUtil._
 
 class UdpIntegrationSpec extends AkkaSpec("""
     akka.loglevel = INFO
