+++ akka-remote/src/test/scala/akka/remote/RemotingSpec.scala
@@ -17,7 +17,7 @@ import scala.concurrent.Await
 import scala.concurrent.Future
 import scala.concurrent.duration._
 import scala.concurrent.forkjoin.ThreadLocalRandom
-import akka.TestUtils.temporaryServerAddress
+import akka.testkit.SocketUtil.temporaryServerAddress
 
 object RemotingSpec {
